- Implemented more VFPU stuff - More (temporal) try/catch on GUI - More function stubs - Added SafeILGenerator.MacroIfElse

+++ CSPspEmu.Core.Cpu/Emiter/Emiters/CpuEmiter.ArithmeticLogic.cs
@@ -83,7 +83,7 @@ namespace CSPspEmu.Core.Cpu.Emiter
 		public void sltu() { MipsMethodEmiter.OP_3REG_Unsigned(RD, RS, RT, () => { SafeILGenerator.CompareBinary(SafeBinaryComparison.LessThanUnsigned); }); }
 
 		public void slti() { MipsMethodEmiter.OP_2REG_IMM_Signed(RT, RS, (short)Instruction.IMM, () => { SafeILGenerator.CompareBinary(SafeBinaryComparison.LessThanSigned); }); }
-		public void sltiu() { MipsMethodEmiter.OP_2REG_IMM_Unsigned(RT, RS, (uint)Instruction.IMM, () => { SafeILGenerator.CompareBinary(SafeBinaryComparison.LessThanUnsigned); }); }
+		public void sltiu() { MipsMethodEmiter.OP_2REG_IMM_Unsigned(RT, RS, (ushort)Instruction.IMM, () => { SafeILGenerator.CompareBinary(SafeBinaryComparison.LessThanUnsigned); }); }
 		//Console.WriteLine("SLTIU: {0} : {1}", Instruction.IMM, (uint)Instruction.IMM);
 
 		/////////////////////////////////////////////////////////////////////////////////////////////////
